refactor: model catalog refactored and optimised
