+++ transports/bifrost-http/lib/config_test.go
@@ -953,7 +953,12 @@ func (m *MockConfigStore) DeletePricingOverride(ctx context.Context, id string, 
 }
 
 // Model parameters
-func (m *MockConfigStore) GetModelParameters(ctx context.Context, model string) (*tables.TableModelParameters, error) {
+
+func (m *MockConfigStore) GetModelParameters(ctx context.Context) ([]tables.TableModelParameters, error) {
+	return nil, nil
+}
+
+func (m *MockConfigStore) GetModelParametersByModel(ctx context.Context, model string) (*tables.TableModelParameters, error) {
 	return nil, nil
 }
 
@@ -4863,10 +4868,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 	// === Scenario 1: Azure config in DB + same in file -> hash matches, no update ===
 	t.Run("SameAzureConfig_NoUpdate", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -4875,10 +4880,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -4898,10 +4903,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 	// === Scenario 2: Azure config in DB + different endpoint in file -> hash differs ===
 	t.Run("DifferentEndpoint_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -4910,10 +4915,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://different-azure.openai.azure.com"), // Changed!
@@ -4933,10 +4938,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 	// === Scenario 3: Azure config in DB + different APIVersion in file -> hash differs ===
 	t.Run("DifferentAPIVersion_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -4945,10 +4950,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -4968,10 +4973,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 	// === Scenario 4: Azure config in DB + different Deployments map in file -> hash differs ===
 	t.Run("DifferentDeployments_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -4979,10 +4984,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment", "gpt-3.5-turbo": "gpt-35-turbo-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -5065,10 +5070,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 	// === Scenario 7: APIVersion nil vs set -> hash differs ===
 	t.Run("APIVersionNilVsSet_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -5077,10 +5082,10 @@ func TestKeyHashComparison_AzureConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -5103,10 +5108,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 1: Bedrock config in DB + same in file -> hash matches, no update ===
 	t.Run("SameBedrockConfig_NoUpdate", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5116,10 +5121,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5140,10 +5145,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 2: Bedrock config in DB + different AccessKey/SecretKey -> hash differs ===
 	t.Run("DifferentAccessKey_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5153,10 +5158,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAI44QH8DHBEXAMPLE"), // Changed!
@@ -5177,10 +5182,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 3: Bedrock config in DB + different SecretKey -> hash differs ===
 	t.Run("DifferentSecretKey_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5190,10 +5195,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5214,10 +5219,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 4: Bedrock config in DB + different Region -> hash differs ===
 	t.Run("DifferentRegion_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5227,10 +5232,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5251,10 +5256,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 5: Bedrock config in DB + different ARN -> hash differs ===
 	t.Run("DifferentARN_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5265,10 +5270,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5290,10 +5295,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 6: Bedrock config in DB + different Deployments -> hash differs ===
 	t.Run("DifferentDeployments_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5303,11 +5308,11 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"claude-3":   "claude-3-inference-profile", "claude-3.5": "claude-35-inference-profile"},
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile", "claude-3.5": "claude-35-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
 				SecretKey: *schemas.NewEnvVar("wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"),
@@ -5393,10 +5398,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	// === Scenario 9: SessionToken nil vs set -> hash differs ===
 	t.Run("SessionTokenNilVsSet_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5407,10 +5412,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey:    *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5433,10 +5438,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 	t.Run("IAMRoleAuthVsExplicitCredentials_UpdateTriggered", func(t *testing.T) {
 		// IAM role auth: empty AccessKey and SecretKey
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar(""),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar(""),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar(""), // Empty for IAM role auth
@@ -5447,10 +5452,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 
 		// Explicit credentials
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar(""),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar(""),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "claude-3-inference-profile"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5473,10 +5478,10 @@ func TestKeyHashComparison_BedrockConfigSyncScenarios(t *testing.T) {
 func TestProviderHashComparison_AzureProviderFullLifecycle(t *testing.T) {
 	// === STEP 1: Initial state - Azure provider exists in DB from previous config.json ===
 	initialAzureKey := schemas.Key{
-		ID:     "azure-key-1",
-		Name:   "azure-openai-key",
-		Value:  *schemas.NewEnvVar("azure-api-key-initial"),
-		Weight: 1,
+		ID:      "azure-key-1",
+		Name:    "azure-openai-key",
+		Value:   *schemas.NewEnvVar("azure-api-key-initial"),
+		Weight:  1,
 		Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 		AzureKeyConfig: &schemas.AzureKeyConfig{
 			Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -5508,10 +5513,10 @@ func TestProviderHashComparison_AzureProviderFullLifecycle(t *testing.T) {
 	// The key value is edited via dashboard, but the Azure config structure stays the same
 	// Provider config hash should remain unchanged
 	dashboardEditedKey := schemas.Key{
-		ID:     "azure-key-1",
-		Name:   "azure-openai-key",
-		Value:  *schemas.NewEnvVar("azure-api-key-dashboard-edited"), // Changed via dashboard!
-		Weight: 1,
+		ID:      "azure-key-1",
+		Name:    "azure-openai-key",
+		Value:   *schemas.NewEnvVar("azure-api-key-dashboard-edited"), // Changed via dashboard!
+		Weight:  1,
 		Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 		AzureKeyConfig: &schemas.AzureKeyConfig{
 			Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -5541,10 +5546,10 @@ func TestProviderHashComparison_AzureProviderFullLifecycle(t *testing.T) {
 	sameFileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "azure-key-1",
-				Name:   "azure-openai-key",
-				Value:  *schemas.NewEnvVar("azure-api-key-initial"), // Original value from file
-				Weight: 1,
+				ID:      "azure-key-1",
+				Name:    "azure-openai-key",
+				Value:   *schemas.NewEnvVar("azure-api-key-initial"), // Original value from file
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 				AzureKeyConfig: &schemas.AzureKeyConfig{
 					Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -5578,11 +5583,11 @@ func TestProviderHashComparison_AzureProviderFullLifecycle(t *testing.T) {
 	newFileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "azure-key-1",
-				Name:   "azure-openai-key",
-				Value:  *schemas.NewEnvVar("azure-api-key-initial"),
-				Weight: 1,
-				Aliases: schemas.KeyAliases{"gpt-4":  "gpt-4-deployment", "gpt-4o": "gpt-4o-deployment"},
+				ID:      "azure-key-1",
+				Name:    "azure-openai-key",
+				Value:   *schemas.NewEnvVar("azure-api-key-initial"),
+				Weight:  1,
+				Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment", "gpt-4o": "gpt-4o-deployment"},
 				AzureKeyConfig: &schemas.AzureKeyConfig{
 					Endpoint:   *schemas.NewEnvVar("https://new-azure.openai.azure.com"), // Changed!
 					APIVersion: schemas.NewEnvVar("2024-10-21"),                          // Changed!
@@ -5688,10 +5693,10 @@ func TestProviderHashComparison_AzureProviderFullLifecycle(t *testing.T) {
 func TestProviderHashComparison_BedrockProviderFullLifecycle(t *testing.T) {
 	// === STEP 1: Initial state - Bedrock provider exists in DB from previous config.json ===
 	initialBedrockKey := schemas.Key{
-		ID:     "bedrock-key-1",
-		Name:   "aws-bedrock-key",
-		Value:  *schemas.NewEnvVar(""), // Empty for Bedrock with IAM or AccessKey auth
-		Weight: 1,
+		ID:      "bedrock-key-1",
+		Name:    "aws-bedrock-key",
+		Value:   *schemas.NewEnvVar(""), // Empty for Bedrock with IAM or AccessKey auth
+		Weight:  1,
 		Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0"},
 		BedrockKeyConfig: &schemas.BedrockKeyConfig{
 			AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5723,10 +5728,10 @@ func TestProviderHashComparison_BedrockProviderFullLifecycle(t *testing.T) {
 
 	// === STEP 2: Dashboard adds a second key ===
 	dashboardAddedKey := schemas.Key{
-		ID:     "bedrock-key-2",
-		Name:   "aws-bedrock-key-eu",
-		Value:  *schemas.NewEnvVar(""),
-		Weight: 1,
+		ID:      "bedrock-key-2",
+		Name:    "aws-bedrock-key-eu",
+		Value:   *schemas.NewEnvVar(""),
+		Weight:  1,
 		Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0"},
 		BedrockKeyConfig: &schemas.BedrockKeyConfig{
 			AccessKey: *schemas.NewEnvVar("AKIAI44QH8DHBEXAMPLE"),
@@ -5750,10 +5755,10 @@ func TestProviderHashComparison_BedrockProviderFullLifecycle(t *testing.T) {
 	sameFileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -5789,11 +5794,11 @@ func TestProviderHashComparison_BedrockProviderFullLifecycle(t *testing.T) {
 	newFileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
-				Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0", "claude-3-opus":   "anthropic.claude-3-opus-20240229-v1:0"},
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
+				Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0", "claude-3-opus": "anthropic.claude-3-opus-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
 					SecretKey: *schemas.NewEnvVar("wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"),
@@ -5917,11 +5922,11 @@ func TestProviderHashComparison_BedrockProviderFullLifecycle(t *testing.T) {
 	sameNewFileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
-				Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0", "claude-3-opus":   "anthropic.claude-3-opus-20240229-v1:0"},
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
+				Aliases: schemas.KeyAliases{"claude-3-sonnet": "anthropic.claude-3-sonnet-20240229-v1:0", "claude-3-opus": "anthropic.claude-3-opus-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
 					SecretKey: *schemas.NewEnvVar("wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"),
@@ -5960,10 +5965,10 @@ func TestProviderHashComparison_AzureNewProviderFromConfig(t *testing.T) {
 	fileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "azure-key-1",
-				Name:   "azure-openai-key",
-				Value:  *schemas.NewEnvVar("azure-api-key-123"),
-				Weight: 1,
+				ID:      "azure-key-1",
+				Name:    "azure-openai-key",
+				Value:   *schemas.NewEnvVar("azure-api-key-123"),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 				AzureKeyConfig: &schemas.AzureKeyConfig{
 					Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -6027,10 +6032,10 @@ func TestProviderHashComparison_BedrockNewProviderFromConfig(t *testing.T) {
 	fileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"claude-3": "anthropic.claude-3-sonnet-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -6096,10 +6101,10 @@ func TestProviderHashComparison_AzureDBValuePreservedWhenHashMatches(t *testing.
 	dbConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "azure-key-1",
-				Name:   "azure-openai-key",
-				Value:  *schemas.NewEnvVar("DASHBOARD-EDITED-SECRET-KEY"), // Dashboard edited this!
-				Weight: 1,
+				ID:      "azure-key-1",
+				Name:    "azure-openai-key",
+				Value:   *schemas.NewEnvVar("DASHBOARD-EDITED-SECRET-KEY"), // Dashboard edited this!
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 				AzureKeyConfig: &schemas.AzureKeyConfig{
 					Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -6125,10 +6130,10 @@ func TestProviderHashComparison_AzureDBValuePreservedWhenHashMatches(t *testing.
 	fileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "azure-key-1",
-				Name:   "azure-openai-key",
-				Value:  *schemas.NewEnvVar("original-key-from-file"), // Different value than DB!
-				Weight: 1,
+				ID:      "azure-key-1",
+				Name:    "azure-openai-key",
+				Value:   *schemas.NewEnvVar("original-key-from-file"), // Different value than DB!
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 				AzureKeyConfig: &schemas.AzureKeyConfig{
 					Endpoint:   *schemas.NewEnvVar("https://myazure.openai.azure.com"), // Same
@@ -6182,10 +6187,10 @@ func TestProviderHashComparison_BedrockDBValuePreservedWhenHashMatches(t *testin
 	dbConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"claude-3": "anthropic.claude-3-sonnet-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("DASHBOARD-EDITED-ACCESS-KEY"), // Dashboard edited!
@@ -6212,10 +6217,10 @@ func TestProviderHashComparison_BedrockDBValuePreservedWhenHashMatches(t *testin
 	fileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"claude-3": "anthropic.claude-3-sonnet-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),                     // Different!
@@ -6301,10 +6306,10 @@ func TestProviderHashComparison_AzureConfigChangedInFile(t *testing.T) {
 	fileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "azure-key-1",
-				Name:   "azure-openai-key",
-				Value:  *schemas.NewEnvVar("azure-api-key-123"),
-				Weight: 1,
+				ID:      "azure-key-1",
+				Name:    "azure-openai-key",
+				Value:   *schemas.NewEnvVar("azure-api-key-123"),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"gpt-4o": "gpt-4o-deployment"},
 				AzureKeyConfig: &schemas.AzureKeyConfig{
 					Endpoint:   *schemas.NewEnvVar("https://NEW-azure.openai.azure.com"), // Changed!
@@ -6389,10 +6394,10 @@ func TestProviderHashComparison_BedrockConfigChangedInFile(t *testing.T) {
 	fileConfig := configstore.ProviderConfig{
 		Keys: []schemas.Key{
 			{
-				ID:     "bedrock-key-1",
-				Name:   "aws-bedrock-key",
-				Value:  *schemas.NewEnvVar(""),
-				Weight: 1,
+				ID:      "bedrock-key-1",
+				Name:    "aws-bedrock-key",
+				Value:   *schemas.NewEnvVar(""),
+				Weight:  1,
 				Aliases: schemas.KeyAliases{"claude-3-opus": "anthropic.claude-3-opus-20240229-v1:0"},
 				BedrockKeyConfig: &schemas.BedrockKeyConfig{
 					AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -14309,10 +14314,10 @@ func TestKeyHashComparison_VertexConfigSyncScenarios(t *testing.T) {
 	// === Scenario 1: Vertex config in DB + same in file -> hash matches, no update ===
 	t.Run("SameVertexConfig_NoUpdate", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID:       *schemas.NewEnvVar("my-project-123"),
@@ -14323,10 +14328,10 @@ func TestKeyHashComparison_VertexConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID:       *schemas.NewEnvVar("my-project-123"),
@@ -14453,10 +14458,10 @@ func TestKeyHashComparison_VertexConfigSyncScenarios(t *testing.T) {
 	// === Scenario 5: Vertex config in DB + different Deployments map in file -> hash differs ===
 	t.Run("DifferentDeployments_UpdateTriggered", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-api-key-123"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-api-key-123"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project-123"),
@@ -14465,11 +14470,11 @@ func TestKeyHashComparison_VertexConfigSyncScenarios(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-api-key-123"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"gemini-pro":     "gemini-pro-endpoint", "gemini-1.5-pro": "gemini-15-pro-endpoint"},
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-api-key-123"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint", "gemini-1.5-pro": "gemini-15-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project-123"),
 				Region:    *schemas.NewEnvVar("us-central1"),
@@ -14852,10 +14857,10 @@ func TestProviderHashComparison_VertexConfigChangedInFile(t *testing.T) {
 func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 	t.Run("AddDeployment", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -14863,11 +14868,11 @@ func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"gpt-4":  "gpt-4-deployment", "gpt-4o": "gpt-4o-deployment"},
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment", "gpt-4o": "gpt-4o-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
 			},
@@ -14883,21 +14888,21 @@ func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 
 	t.Run("RemoveDeployment", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"gpt-4":  "gpt-4-deployment", "gpt-4o": "gpt-4o-deployment"},
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment", "gpt-4o": "gpt-4o-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
 			},
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -14914,10 +14919,10 @@ func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 
 	t.Run("ModifyDeploymentValue", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment-v1"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -14925,10 +14930,10 @@ func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment-v2"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -14955,10 +14960,10 @@ func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "azure-key",
-			Value:  *schemas.NewEnvVar("azure-api-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "azure-key",
+			Value:   *schemas.NewEnvVar("azure-api-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gpt-4": "gpt-4-deployment"},
 			AzureKeyConfig: &schemas.AzureKeyConfig{
 				Endpoint: *schemas.NewEnvVar("https://myazure.openai.azure.com"),
@@ -14978,10 +14983,10 @@ func TestKeyHashComparison_AzureDeploymentsChange(t *testing.T) {
 func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 	t.Run("AddDeployment", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "arn:aws:bedrock:us-east-1::inference-profile/claude-3"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -14991,11 +14996,11 @@ func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-key"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"claude-3":   "arn:aws:bedrock:us-east-1::inference-profile/claude-3", "claude-3.5": "arn:aws:bedrock:us-east-1::inference-profile/claude-3.5"},
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-key"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"claude-3": "arn:aws:bedrock:us-east-1::inference-profile/claude-3", "claude-3.5": "arn:aws:bedrock:us-east-1::inference-profile/claude-3.5"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
 				SecretKey: *schemas.NewEnvVar("wJalrXUtnFEMI"),
@@ -15013,11 +15018,11 @@ func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 
 	t.Run("RemoveDeployment", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-key"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"claude-3":   "arn:aws:bedrock:us-east-1::inference-profile/claude-3", "claude-3.5": "arn:aws:bedrock:us-east-1::inference-profile/claude-3.5"},
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-key"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"claude-3": "arn:aws:bedrock:us-east-1::inference-profile/claude-3", "claude-3.5": "arn:aws:bedrock:us-east-1::inference-profile/claude-3.5"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
 				SecretKey: *schemas.NewEnvVar("wJalrXUtnFEMI"),
@@ -15026,10 +15031,10 @@ func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "arn:aws:bedrock:us-east-1::inference-profile/claude-3"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -15048,10 +15053,10 @@ func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 
 	t.Run("ModifyDeploymentValue", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "arn:aws:bedrock:us-east-1::inference-profile/claude-3-old"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -15061,10 +15066,10 @@ func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "bedrock-key",
-			Value:  *schemas.NewEnvVar("bedrock-key"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "bedrock-key",
+			Value:   *schemas.NewEnvVar("bedrock-key"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"claude-3": "arn:aws:bedrock:us-east-1::inference-profile/claude-3-new"},
 			BedrockKeyConfig: &schemas.BedrockKeyConfig{
 				AccessKey: *schemas.NewEnvVar("AKIAIOSFODNN7EXAMPLE"),
@@ -15086,10 +15091,10 @@ func TestKeyHashComparison_BedrockDeploymentsChange(t *testing.T) {
 func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 	t.Run("AddDeployment", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
@@ -15098,11 +15103,11 @@ func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"gemini-pro":     "gemini-pro-endpoint", "gemini-1.5-pro": "gemini-15-pro-endpoint"},
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint", "gemini-1.5-pro": "gemini-15-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
 				Region:    *schemas.NewEnvVar("us-central1"),
@@ -15119,11 +15124,11 @@ func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 
 	t.Run("RemoveDeployment", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
-			Aliases: schemas.KeyAliases{"gemini-pro":     "gemini-pro-endpoint", "gemini-1.5-pro": "gemini-15-pro-endpoint"},
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
+			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint", "gemini-1.5-pro": "gemini-15-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
 				Region:    *schemas.NewEnvVar("us-central1"),
@@ -15131,10 +15136,10 @@ func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
@@ -15152,10 +15157,10 @@ func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 
 	t.Run("ModifyDeploymentValue", func(t *testing.T) {
 		dbKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint-v1"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
@@ -15164,10 +15169,10 @@ func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint-v2"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
@@ -15196,10 +15201,10 @@ func TestKeyHashComparison_VertexDeploymentsChange(t *testing.T) {
 		}
 
 		fileKey := schemas.Key{
-			ID:     "key-1",
-			Name:   "vertex-key",
-			Value:  *schemas.NewEnvVar("vertex-creds"),
-			Weight: 1,
+			ID:      "key-1",
+			Name:    "vertex-key",
+			Value:   *schemas.NewEnvVar("vertex-creds"),
+			Weight:  1,
 			Aliases: schemas.KeyAliases{"gemini-pro": "gemini-pro-endpoint"},
 			VertexKeyConfig: &schemas.VertexKeyConfig{
 				ProjectID: *schemas.NewEnvVar("my-project"),
@@ -15703,7 +15708,7 @@ func TestConfigSchemaSyncTopLevel(t *testing.T) {
 
 func TestResolveFrameworkPricingConfig(t *testing.T) {
 	defaultURL := modelcatalog.DefaultPricingURL
-	defaultSyncSeconds := int64(modelcatalog.DefaultPricingSyncInterval.Seconds())
+	defaultSyncSeconds := int64(modelcatalog.DefaultSyncInterval.Seconds())
 	fileURL := "https://example.com/pricing.json"
 	fileSyncDuration := 12 * time.Hour
 	fileSyncSeconds := int64(fileSyncDuration.Seconds())
@@ -15760,7 +15765,7 @@ func TestResolveFrameworkPricingConfig(t *testing.T) {
 		require.Equal(t, defaultURL, *normalizedTable.PricingURL)
 		require.Equal(t, defaultSyncSeconds, *normalizedTable.PricingSyncInterval)
 		require.Equal(t, defaultURL, *normalizedModelCatalog.PricingURL)
-		require.Equal(t, modelcatalog.DefaultPricingSyncInterval, *normalizedModelCatalog.PricingSyncInterval)
+		require.Equal(t, modelcatalog.DefaultSyncInterval, *normalizedModelCatalog.PricingSyncInterval)
 	})
 
 	t.Run("invalid db interval falls back and requests db update", func(t *testing.T) {
@@ -15776,7 +15781,7 @@ func TestResolveFrameworkPricingConfig(t *testing.T) {
 		require.Equal(t, dbURL, *normalizedTable.PricingURL)
 		require.Equal(t, defaultSyncSeconds, *normalizedTable.PricingSyncInterval)
 		require.Equal(t, dbURL, *normalizedModelCatalog.PricingURL)
-		require.Equal(t, modelcatalog.DefaultPricingSyncInterval, *normalizedModelCatalog.PricingSyncInterval)
+		require.Equal(t, modelcatalog.DefaultSyncInterval, *normalizedModelCatalog.PricingSyncInterval)
 	})
 }
